쿠키를 이용한 로그인, 로그아웃 처리

+++ src/main/java/com/sbs/basic1/boundedContext/member/entity/Member.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 @Getter
 public class Member {
-  private static int lastId;
+  private static long lastId;
   private final long id;
   private final String username;
   @Setter
